requirements.txt
aiogram==3.3.0
openai==1.10.0
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.9
itsdangerous==2.1.2
